+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@layerfi/components",
-  "version": "0.1.58",
+  "version": "0.1.59",
   "description": "Layer React Components",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
