+++ src/components/ProfitAndLossReport/index.tsx
@@ -1,0 +1,1 @@
+export { ProfitAndLossReport } from './ProfitAndLossReport'
