+++ src/views/Reports/Reports.tsx
@@ -1,14 +1,8 @@
-import React, { RefObject, useContext, useRef, useState } from 'react'
-import { Layer } from '../../api/layer'
+import React, { RefObject, useRef, useState } from 'react'
 import { BalanceSheet } from '../../components/BalanceSheet'
 import { BalanceSheetStringOverrides } from '../../components/BalanceSheet/BalanceSheet'
-import { Button, ButtonVariant, RetryButton } from '../../components/Button'
 import { Container } from '../../components/Container'
-import {
-  DatePickerMode,
-  DateRangeDatePickerModes,
-} from '../../components/DatePicker/DatePicker'
-import { Panel } from '../../components/Panel'
+import { DateRangeDatePickerModes } from '../../components/DatePicker/DatePicker'
 import { ProfitAndLoss } from '../../components/ProfitAndLoss'
 import { ProfitAndLossCompareOptionsProps } from '../../components/ProfitAndLossCompareOptions/ProfitAndLossCompareOptions'
 import { ProfitAndLossDetailedChartsStringOverrides } from '../../components/ProfitAndLossDetailedCharts/ProfitAndLossDetailedCharts'
@@ -17,11 +11,10 @@ import { StatementOfCashFlow } from '../../components/StatementOfCashFlow'
 import { StatementOfCashFlowStringOverrides } from '../../components/StatementOfCashFlow/StatementOfCashFlow'
 import { Toggle } from '../../components/Toggle'
 import { View } from '../../components/View'
-import { useLayerContext } from '../../contexts/LayerContext'
-import DownloadCloud from '../../icons/DownloadCloud'
+import { DownloadButtonStringOverrides } from '../../components/ProfitAndLossReport/ProfitAndLossReport'
 import { MoneyFormat } from '../../types'
 
-interface ReportsStringOverrides {
+export interface ReportsStringOverrides {
   title?: string
   downloadButton?: DownloadButtonStringOverrides
   profitAndLoss?: {
@@ -60,71 +53,6 @@ export interface ReportsPanelProps {
   statementOfCashFlowConfig?: {
     datePickerMode?: DateRangeDatePickerModes
   }
-}
-
-interface DownloadButtonStringOverrides {
-  downloadButtonText?: string
-  retryButtonText?: string
-}
-
-interface DownloadButtonProps {
-  stringOverrides?: DownloadButtonStringOverrides
-  moneyFormat?: MoneyFormat
-}
-
-const DownloadButton = ({
-  stringOverrides,
-  moneyFormat,
-}: DownloadButtonProps) => {
-  const { dateRange } = useContext(ProfitAndLoss.Context)
-  const { auth, businessId, apiUrl } = useLayerContext()
-  const [requestFailed, setRequestFailed] = useState(false)
-
-  const handleClick = async () => {
-    const month = (dateRange.startDate.getMonth() + 1).toString()
-    const year = dateRange.startDate.getFullYear().toString()
-    const getProfitAndLossCsv = Layer.getProfitAndLossCsv(
-      apiUrl,
-      auth.access_token,
-      {
-        params: {
-          businessId: businessId,
-          year: year,
-          month: month,
-          moneyFormat: moneyFormat,
-        },
-      },
-    )
-    try {
-      const result = await getProfitAndLossCsv()
-      if (result?.data?.presignedUrl) {
-        window.location.href = result.data.presignedUrl
-        setRequestFailed(false)
-      } else {
-        setRequestFailed(true)
-      }
-    } catch (e) {
-      setRequestFailed(true)
-    }
-  }
-
-  return requestFailed ? (
-    <RetryButton
-      onClick={handleClick}
-      className='Layer__download-retry-btn'
-      error={'Approval failed. Check connection and retry in few seconds.'}
-    >
-      {stringOverrides?.retryButtonText || 'Retry'}
-    </RetryButton>
-  ) : (
-    <Button
-      variant={ButtonVariant.secondary}
-      rightIcon={<DownloadCloud size={12} />}
-      onClick={handleClick}
-    >
-      {stringOverrides?.downloadButtonText || 'Download'}
-    </Button>
-  )
 }
 
 const getOptions = (enabledReports: ReportType[]) => {
@@ -203,48 +131,16 @@ const ReportsPanel = ({
   profitAndLossConfig,
   statementOfCashFlowConfig,
 }: ReportsPanelProps) => {
-  const { sidebarScope } = useContext(ProfitAndLoss.Context)
   return (
     <>
       {openReport === 'profitAndLoss' && (
-        <View
-          type='panel'
-          headerControls={
-            <>
-              <ProfitAndLoss.DatePicker
-                datePickerMode={profitAndLossConfig?.datePickerMode}
-              />
-              <div className='Layer__compare__controls__wrapper'>
-                {comparisonConfig && (
-                  <ProfitAndLoss.CompareOptions
-                    tagComparisonOptions={comparisonConfig.tagComparisonOptions}
-                    defaultTagFilter={comparisonConfig.defaultTagFilter}
-                  />
-                )}
-                <DownloadButton
-                  stringOverrides={stringOverrides?.downloadButton}
-                  moneyFormat={profitAndLossConfig?.csvMoneyFormat}
-                />
-              </div>
-            </>
-          }
-        >
-          <Panel
-            sidebar={
-              <ProfitAndLoss.DetailedCharts
-                showDatePicker={false}
-                stringOverrides={stringOverrides?.profitAndLoss?.detailedCharts}
-              />
-            }
-            sidebarIsOpen={Boolean(sidebarScope)}
-            parentRef={containerRef}
-          >
-            <ProfitAndLoss.Table
-              asContainer={false}
-              stringOverrides={stringOverrides?.profitAndLoss?.table}
-            />
-          </Panel>
-        </View>
+        <ProfitAndLoss.Report
+          stringOverrides={stringOverrides}
+          comparisonConfig={comparisonConfig}
+          datePickerMode={profitAndLossConfig?.datePickerMode}
+          csvMoneyFormat={profitAndLossConfig?.csvMoneyFormat}
+          parentRef={containerRef}
+        />
       )}
       {openReport === 'balanceSheet' && (
         <BalanceSheet stringOverrides={stringOverrides?.balanceSheet} />
